backend/requirements.txt
django==3.0.5
djangorestframework
psycopg2-binary
django-cors-headers
djangorestframework-simplejwt==4.4.0
djoser
PyJWT==1.7.1
sqlparse==0.4.1
